All search functions moved to one class
