--- src/main/java/edu/matc/rest/ResponseHolidayAll.java
@@ -1,33 +1,0 @@
-package edu.matc.rest;
-
-import com.google.gson.Gson;
-import edu.matc.entity.Holiday;
-import edu.matc.persistence.GenericDao;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Response;
-import java.util.List;
-
-@Path("/allHolidays")
-public class ResponseHolidayAll {
-    private final Logger logger = LogManager.getLogger(this.getClass());
-
-    // The Java method will process HTTP GET requests
-    @GET
-    @Produces("text/json")
-    public Response getMessage() {
-        GenericDao<Holiday> holidayDao = new GenericDao<>(Holiday.class);
-        List<Holiday> holidays = holidayDao.getAll();
-        logger.debug("The list of holidays:" + holidays);
-
-        // Convert the list of Holiday objects to JSON using Gson
-        Gson gson = new Gson();
-        String json = gson.toJson(holidays);
-
-        return Response.status(200).entity(json).build();
-    }
-}
